requirements.txt
# SARA - Su Altı Roket Aracı - Python Gereksinimleri
# Raspberry Pi 5 uyumlu kütüphaneler

# ---- GPIO Kütüphanesi (Pi 5 uyumlu) ----
# RPi.GPIO yerine rpi-lgpio kullan
rpi-lgpio>=0.4.0

# ---- MAVLink ----
pymavlink>=2.4.37

# ---- Sensör Kütüphaneleri ----
smbus2>=0.4.2              # I2C haberleşmesi

# ---- Temel Kütüphaneler ----
numpy>=1.21.0              # Matematik işlemleri
scipy>=1.7.0               # Bilimsel hesaplamalar

# ---- Logging ve Utilities ----
colorlog>=6.7.0            # Renkli log çıktıları

# ---- Test ve Debugging ----
pytest>=7.0.0              # Unit testler için

# ---- Alternatif GPIO (opsiyonel) ----
# pigpio>=1.78               # pigpio daemon tabanlı (alternatif)
# gpiozero>=1.6.2            # High-level GPIO (opsiyonel)

# ---- Sistem İzleme ----
psutil>=5.9.0              # Sistem kaynakları

# ---- Config ve Utilities ----
pyyaml>=6.0                # YAML config dosyaları (gelecekte)
python-dateutil>=2.8.2    # Tarih işlemleri
